nautilus/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.9.0
gunicorn==20.1.0
importlib-metadata==6.0.0
mysqlclient==2.1.0
numpy==1.24.2
pandas==1.5.3
pipenv==2023.2.18
platformdirs==3.1.0
pytz==2022.7.1
sqlparse==0.4.2
typing_extensions==4.5.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
yfinance==0.2.12
zipp==3.15.0
